fix(mf2): Rename pattern/ -> expression/
