+++ packages/mf2-messageformat/src/expression/function-annotation.test.ts
