+++ packages/mf2-messageformat/src/expression/literal.ts
