+++ packages/mf2-messageformat/src/expression/unsupported-annotation.test.ts
