+++ packages/mf2-messageformat/src/expression/value.ts
