+++ packages/mf2-messageformat/src/expression/variable-ref.ts
