+++ packages/mf2-messageformat/src/expression/variable.test.ts
