+++ packages/mf2-messageformat/src/index.ts
@@ -6,8 +6,8 @@ export {
   MessageResolutionError,
   MessageSyntaxError
 } from './errors.js';
+export * from './expression/index.js';
 export { MessageFormat, MessageFormatOptions } from './messageformat.js';
 export { stringifyMessage } from './stringifier/message.js';
-export * from './pattern/index.js';
 export * from './runtime/index.js';
 export { validate } from './extra/validate.js'; // must be after ./messageformat -- but why!?
